requirements.txt
onnxruntime>=1.14.1
numpy>=1.24.3
pytest>=7.3.1
uvicorn[standard]>=0.22.0
requests>=2.31.0
pillow>=9.5.0
fastapi>=0.95.2